Changed routes, added observers, modified migrations

+++ database/migrations/2021_10_24_114133_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->string('order_id')->unique();
-            $table->float('total');
+            $table->decimal('total');
             $table->timestamp('date');
             $table->foreignId('status_id')->constrained('order_statuses')->restrictOnDelete();
             $table->timestamps();
